DIP-0026: fix -Wextra warning in update_registrar version inherit
The version-inherit ternary mixed a uint16_t (std::min result) with the
ProTxVersion enum, which g++ flags under -Wextra (enumerated and non-enumerated
type in conditional expression) and would fail an -Werror CI build. Cast the
enum branch to uint16_t.

g++-only: clang did not emit it. Swept the whole DIP-0026 diff on g++ and
confirmed this was the only warning; the rest is warning-clean.

+++ src/rpc/evo.cpp
@@ -1236,7 +1236,7 @@ static RPCHelpMan protx_update_registrar_wrapper(const bool specific_legacy_bls_
         ptx.payoutShares = dmn->pdmnState->payoutShares;
         ptx.nVersion = dmn->pdmnState->payoutShares.empty()
                            ? std::min<uint16_t>(ptx.nVersion, ProTxVersion::BasicBLS)
-                           : ProTxVersion::MultiPayout;
+                           : static_cast<uint16_t>(ProTxVersion::MultiPayout);
     }
     // representative payout dest for the fee-source default (single address, or the first share)
     CTxDestination payoutDest;
